requirements.txt
psycopg2==2.6.1
celery==3.1.23
redis==2.10.5
pymongo==3.2.2
djangorestframework==3.3.3
Scrapy==1.1
selenium==2.53.2
mongoengine==0.10.6